feat: add support for generating commit messages from the last commit 🚀
This commit adds a new feature to the script that allows generating commit messages based on the last commit instead of just staged changes. The new `--source` option can be set to either 'diff' (default, for staged changes) or 'last-commit'. When `--source last-commit` is used, the script fetches the diff between the current HEAD and its parent commit to generate the commit message. This is useful for scenarios where users want to review or modify their last commit before amending it.