chore: add alias for commit with all changes 🛠
This commit adds a new Git alias `cia` to the `.gitconfig` file. The alias stands for "commit all" and allows users to commit all changes in their working directory with the command `git cia`. This shortcut simplifies the process of committing all modifications, making it more efficient for developers who frequently need to commit changes.