feat: add new tmuxp configuration for ticketscloud 🚀
This commit introduces a new `.tmuxp` configuration file named `ticketscloud.yml`. The configuration sets up a session for the `ticketscloud` project with a single window focused on a common directory. Additionally, it updates the `dotfiles.yml` to include a new window named `ghostty`, which starts an `nvim` session for editing the Ghostty configuration file.